Extract Dijkstra path finder
